+++ lib/mongoid/shell/commands/mongostat.rb
@@ -6,6 +6,8 @@ module Mongoid
         include Mongoid::Shell::Properties::Database
         include Mongoid::Shell::Properties::Username
         include Mongoid::Shell::Properties::Password
+        include Mongoid::Shell::Properties::AuthenticationDatabase
+        include Mongoid::Shell::Properties::SSL
 
         option :host
         option :username
@@ -15,6 +17,8 @@ module Mongoid
         option :noheaders
         option :http
         option :all
+        option :authenticationDatabase
+        option :ssl
       end
     end
   end