+++ lib/mongoid/shell/properties/authentication_database.rb
@@ -1,0 +1,13 @@
+module Mongoid
+  module Shell
+    module Properties
+      module AuthenticationDatabase
+        attr_accessor :authenticationDatabase
+
+        def authenticationDatabase
+          @authenticationDatabase || session.options['auth_source']
+        end
+      end
+    end
+  end
+end