+++ lib/mongoid/shell/properties/ssl.rb
@@ -1,0 +1,14 @@
+module Mongoid
+  module Shell
+    module Properties
+      module SSL
+        attr_accessor :ssl
+
+        def ssl
+          return @ssl unless @ssl.nil?
+          session.options['ssl']
+        end
+      end
+    end
+  end
+end