+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 0.4.6 (Next)
 
 * Your contribution here.
+* [#16](https://github.com/mongoid/mongoid-shell/pull/16): Add `--ssl` and `--authenticationDatabase` support to all commands - [@cavvia](https://github.com/cavvia).
 
 ### 0.4.5 (9/7/2017)